Allow the dispatcher concurrency to be overridden per channel (#1669)
* Change consumer dispatch concurrency value to a `ushort`

* Create `CreateChannelAsync` extension to use `DefaultConsumerDispatchConcurrency`

* Standardize on `ConsumerDispatchConcurrency` name

+++ projects/Benchmarks/ConsumerDispatching/ConsumerDispatcher.cs
@@ -35,7 +35,7 @@ namespace RabbitMQ.Benchmarks
         public int Count { get; set; }
 
         [Params(1, 2)]
-        public int Concurrency { get; set; }
+        public ushort Concurrency { get; set; }
 
         [GlobalSetup(Target = nameof(AsyncConsumerDispatcher))]
         public async Task SetUpAsyncConsumer()
